pyproject.toml
[tool.poetry]
name = "hcc-pipeline"
version = "0.1.0"
description = "HCC Documentation Automation Pipeline"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
google-generativeai = "^0.3.0"
pandas = "^2.0.3"
python-dotenv = "^1.0.0"
uvicorn = "^0.34.0"
fastapi = "^0.115.11"
langgraph = {extras = ["cli"], version = "^0.3.10"}
langgraph-cli = "^0.1.76"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
pytest-mock = "^3.11.1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"